Some big changes are being introduced on this commit.
The first is related on the way that Akelos bootstraps.

Before this commit the file config/boot.php contained most of Akelos environment guessing code. It was the place were constants were defined.

That made users have to update their config/boot.php file every time the file was updated which was too frequently.

From now ahead constants are defined at lib/constants.php which is in the framework path instead of in the application path.

config/boot.php will just lookup for the framework and base dir and that will hardly change at all.

The other change is related with the reduction of writable paths in Akelos.
The default location of the cache directory has moved to the newly created ./tmp/cache directory.

Temporary data now resides at ./tmp instead of previous ./cache/tmp


git-svn-id: http://svn.akelos.org/trunk@296 a2fa5c27-f921-0410-a72c-bf682d381be0